% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_find_script.R
\name{internal_find_script}
\alias{internal_find_script}
\title{Internal Function for pathing of files that will be provided}
\usage{
internal_find_script(script_name, package = "dsAnalysis")
}
\arguments{
\item{script_name}{specifies which script is called.}

\item{package}{is the name of the package where the files are included ("dsAnalysis").}
}
\value{
the path of the script_name
}
\description{
This is an internal function which will find the correct path for files.
}
\details{
This function adapts the code from the internal function find_template from the `usethis::use_template` function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
