% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_MockData.R
\name{update_MockData}
\alias{update_MockData}
\title{Function to update the 01_DSLite_Setup.R file}
\usage{
update_MockData(folder_name = NULL, table_names = NULL)
}
\arguments{
\item{folder_name}{the folder where the mock data is stored}

\item{table_names}{optional - given new table names}
}
\value{
adjusted R Script
}
\description{
XXX
}
\details{
XXX.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
