% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dsPackage.R
\name{add_dsPackage}
\alias{add_dsPackage}
\title{Function to add dsPackages to the 01_DSLite_Setup.R file ABCDDE}
\usage{
add_dsPackage(dsPackage = NULL, client = NULL)
}
\arguments{
\item{dsPackage}{name or names of the DataSHIELD server-side packages to add to DSLite instance}

\item{client}{name or names of the matching client-side packages}
}
\value{
adjusted R Script
}
\description{
XXX
}
\details{
XXXXXXsssddasdsdfsdasdfsaasdsadfsdfsdf
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
