% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_project.R
\name{check_project}
\alias{check_project}
\title{Checks that the installed packages match the project's renv.lock}
\usage{
check_project(fix = NULL)
}
\arguments{
\item{fix}{"restore", "snapshot", or NULL to ask}
}
\description{
Checks that the installed packages match the project's renv.lock
}
