% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMockData.R
\name{initMockdata}
\alias{initMockdata}
\title{Function to get mock data from connected data sources}
\usage{
initMockdata(folder_name = NULL, df = "D", datasources = NULL)
}
\arguments{
\item{folder_name}{folder where the mock data shall be stored}

\item{df}{data.frame object on the server-side to produce mock data from}

\item{datasources}{a list of \code{\link[DSI]{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified the default set of connections will be
used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
mock data in the utils/mock_data folder
}
\description{
The function will create a new DataSHIELD analysis project environment
}
\details{
This function adaptssome things.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
