% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMockData.R
\name{initMockdata}
\alias{initMockdata}
\title{Function to get mock data from connected data sources}
\usage{
initMockdata(conns = conns)
}
\arguments{
\item{conns}{specifies the client-side object from DS Login}
}
\value{
mock data in the utils/mock_data folder
}
\description{
The function will create a new DataSHIELD analysis project environment
}
\details{
This function adaptssome things.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
