% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProject.R
\name{initProject}
\alias{initProject}
\title{Initiate a new project environment dedicated for DataSHIELD analysis}
\usage{
initProject(path = "home", name = NULL, switch_to_proj = FALSE)
}
\arguments{
\item{path}{specifies the path where the project shall be initiated. Defaults to the home folder.}

\item{name}{name of the new project}
}
\value{
R Project (not initiated)
}
\description{
The function will create a new DataSHIELD analysis project environment
}
\details{
This function creates a new DataSHIELD analysis project environment that has multiple features:
renv is being used to control R versions for analysis, the .Renviron file in combination with the config.yml files are
used to switch between a realworld setting connecting to live DataSHIELD servers and a local DSLite environment where testing
can be done. Folder structure and some initial R scripts are provided as well to assist the start of DataSHIELD analysis.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
