% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProject.R
\name{initProject}
\alias{initProject}
\title{Function to initiate a new project}
\usage{
initProject(location = NULL, initiate_git = FALSE)
}
\arguments{
\item{location}{specifies the location where the project shall be installed}

\item{initiate_git}{XXXX}
}
\value{
the path of the script_name
}
\description{
The function will create a new DataSHIELD analysis project environment
}
\details{
This function adaptssome things.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
