% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dsPackage.R
\name{install_dsPackage}
\alias{install_dsPackage}
\title{Installs DataSHIELD packages and adds them to the project}
\usage{
install_dsPackage(
  dsPackage = NULL,
  version = NULL,
  ref = NULL,
  source = NULL,
  client = NULL,
  client_version = NULL,
  client_source = NULL
)
}
\arguments{
\item{dsPackage}{name or names of the DataSHIELD server-side packages}

\item{version}{version of the server package}

\item{ref}{GitHub branch, tag or commit of the server package}

\item{source}{GitHub repository of the server package ("owner/repo")}

\item{client}{name of the client package}

\item{client_version}{version of the client package}

\item{client_source}{GitHub repository of the client package ("owner/repo")}
}
\description{
Installs DataSHIELD packages and adds them to the project
}
