% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dsPackage.R
\name{remove_dsPackage}
\alias{remove_dsPackage}
\title{Removes DataSHIELD packages from the DSLite setup and dependencies.R}
\usage{
remove_dsPackage(dsPackage = NULL, client = NULL)
}
\arguments{
\item{dsPackage}{name or names of the DataSHIELD server-side packages to remove}

\item{client}{name or names of the matching client-side packages}
}
\description{
Removes DataSHIELD packages from the DSLite setup and dependencies.R
}
