% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_dsPackages.R
\name{sync_dsPackages}
\alias{sync_dsPackages}
\title{Matches the local packages and DSLite setup to the DataSHIELD servers}
\usage{
sync_dsPackages(conns = NULL, install = TRUE)
}
\arguments{
\item{conns}{DataSHIELD connections; found in the session if not given}

\item{install}{FALSE only reports what would be installed}
}
\description{
Matches the local packages and DSLite setup to the DataSHIELD servers
}
